requirements.txt
google-genai
python-dotenv
yt-dlp
pillow
streamlit
yfinance
pandas
altair